perf(web): isolate catalog presentation bundle
